Clean up obsolete build files
In the past, years ago, the releng project had been used to maintain
references to all RAP Incubator components, and created an aggregated p2
repository.

As of today, the responsibility of the releng project is to provide the
common configuration as a parent pom.xml, and some build scripts that
are used to build the different RAP Incubator components.

Change-Id: I1e29bf673029d124129a1eb81ba53f649904d6d5
Signed-off-by: Markus Knauer <mknauer@eclipsesource.com>